% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactVaRTest-package.R
\docType{package}
\name{ExactVaRTest-package}
\alias{ExactVaRTest}
\alias{ExactVaRTest-package}
\title{ExactVaRTest – Exact Finite-Sample VaR Back-Testing}
\description{
Provides fast dynamic-programming algorithms (C++/Rcpp) – with pure-R
fall-backs – for the exact finite-sample distributions of Christoffersen’s
(1998) Unconditional Coverage (UC), Independence (IND) and Conditional
Coverage (CC) tests.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YourGitHub/ExactVaRTest}
  \item Report bugs at \url{https://github.com/YourGitHub/ExactVaRTest/issues}
}

}
\author{
\strong{Maintainer}: Yujian Chen \email{yujian.chen@polytechnique.edu}

}
