% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{lr_cc_stat}
\alias{lr_cc_stat}
\title{Christoffersen LR_cc statistic}
\usage{
lr_cc_stat(x, alpha = 0.05)
}
\arguments{
\item{x}{0/1 exception series.}

\item{alpha}{Exception probability.}
}
\value{
Numeric LR_cc statistic.
}
\description{
Christoffersen LR_cc statistic
}
