% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{lr_ind_stat}
\alias{lr_ind_stat}
\title{Christoffersen LR_ind statistic}
\usage{
lr_ind_stat(x, alpha = 0.05)
}
\arguments{
\item{x}{0/1 exception series.}

\item{alpha}{Exception probability.}
}
\value{
Numeric LR_ind statistic.
}
\description{
Christoffersen LR_ind statistic
}
\examples{
set.seed(1)
x <- rbinom(50, 1, 0.05)
lr_ind_stat(x, 0.05)
}
