% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{lr_uc_stat}
\alias{lr_uc_stat}
\title{Christoffersen LR_uc statistic}
\usage{
lr_uc_stat(x, alpha = 0.05)
}
\arguments{
\item{x}{0/1 exception series.}

\item{alpha}{Exception probability.}
}
\value{
Numeric LR_uc statistic.
}
\description{
Christoffersen LR_uc statistic
}
