% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{print.ExactVaRBacktest}
\alias{print.ExactVaRBacktest}
\title{Print method for ExactVaRBacktest}
\usage{
\method{print}{ExactVaRBacktest}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{An \code{ExactVaRBacktest} object.}

\item{digits}{Number of digits to print.}

\item{...}{Unused.}
}
\description{
Print method for ExactVaRBacktest
}
