% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{pval_lr_ind}
\alias{pval_lr_ind}
\title{Exact p-value for LR_ind}
\usage{
pval_lr_ind(lr_obs, n, alpha = 0.05, prune_threshold = 1e-15)
}
\arguments{
\item{lr_obs}{Observed LR_ind statistic.}

\item{n}{Sample size.}

\item{alpha}{Exception probability.}

\item{prune_threshold}{State-pruning threshold for DP engine.}
}
\value{
Numeric p-value.
}
\description{
Exact p-value for LR_ind
}
