% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{pval_lr_uc}
\alias{pval_lr_uc}
\title{Exact p-value for LR_uc}
\usage{
pval_lr_uc(lr_obs, n, alpha = 0.05)
}
\arguments{
\item{lr_obs}{Observed LR_uc statistic.}

\item{n}{Sample size.}

\item{alpha}{Exception probability.}
}
\value{
Numeric p-value.
}
\description{
Exact p-value for LR_uc
}
